Describe Compiler Node idea
